Develop

+++ src/app/controllers/rivalriesController.js
@@ -55,6 +55,40 @@ const createRival = async (rival, rivalry) => {
   }
 };
 
+//rivalries of a user
+router.get("/byUser", authMiddleware, async (req, res) => {
+  try {
+    const rivalries = await Rivalry.find({ user: req.userId })
+      .sort({ createdAt: -1 }) // order by most recent
+      .populate([
+        { path: "user", select: "name _id email" },
+        { path: "rivals.rival", select: "name about imageUrl" },
+        { path: "tags", select: "name" },
+      ]);
+
+    return res.send({ rivalries });
+  } catch (err) {
+    res.status(400).send({ error: "Error at listing rivalries" });
+  }
+});
+
+//rivalries that user liked
+router.get("/userLiked", authMiddleware, async (req, res) => {
+  try {
+    const rivalries = await Rivalry.find({ likes: req.userId })
+      .sort({ createdAt: -1 }) // order by most recent
+      .populate([
+        { path: "user", select: "name _id email" },
+        { path: "rivals.rival", select: "name about imageUrl" },
+        { path: "tags", select: "name" },
+      ]);
+
+    return res.send({ rivalries });
+  } catch (err) {
+    res.status(400).send({ error: "Error at listing rivalries" });
+  }
+});
+
 //trending rivalries
 router.get("/trending", async (req, res) => {
   try {
@@ -153,6 +187,10 @@ router.post("/", authMiddleware, async (req, res) => {
             // create the rival if not found
             envolvedRival = await createRival(rival, rivalry);
             await envolvedRival.save();
+          } else {
+            // if rival found push the rivalry to it
+            envolvedRival.rivalries.push(rivalry);
+            await envolvedRival.save();
           }
           rivalry.rivals.push({ rival: envolvedRival });
         }
@@ -195,16 +233,37 @@ router.post("/", authMiddleware, async (req, res) => {
 //update
 router.put("/:rivalryId", authMiddleware, async (req, res) => {
   try {
-    const { title, about, rivals } = req.body;
-    const rivalry = await Rivalry.findByIdAndUpdate(
-      req.params.rivalryId,
-      {
-        title,
-        about,
-      },
-      { new: true }
-    );
+    const { about, tags } = req.body;
+    console.log(req.body);
+    const rivalry = await Rivalry.findById(req.params.rivalryId);
+    rivalry.about = about;
+    let previousTags = rivalry.tags;
+    rivalry.tags = [];
 
+    if (tags !== undefined) {
+      await Promise.all(
+        tags.map(async (tag) => {
+          const envolvedTag = await Tag.findOne({ name: tag });
+          if (envolvedTag) {
+            if (previousTags.includes(envolvedTag)) {
+              // do nothing, already in the array
+            } else {
+              //was not associated, associate with rivalry
+              envolvedTag.rivalries.push(rivalry);
+            }
+            await envolvedTag.save();
+            rivalry.tags.push(envolvedTag);
+          } else {
+            // create the tag and associate
+            const envolvedTag = new Tag({ name: tag });
+
+            envolvedTag.rivalries.push(rivalry);
+            await envolvedTag.save();
+            rivalry.tags.push(envolvedTag);
+          }
+        })
+      );
+    }
     // await Promise.all(
     //   rivals.map(async (rival) => {
     //     // TODO get rival already created and associate if exists
@@ -216,7 +275,16 @@ router.put("/:rivalryId", authMiddleware, async (req, res) => {
     //   })
     // );
     //
-    // await rivalry.save();
+
+    // TODO clean the rivalry references for the tag
+    // previousTags.map((tag) => {
+    //   if (rivalry.tags.filter((t) => t.name === tag.name).length < 1) {
+    //     // check if the tag was dessaociated of that rivalry
+    //     tag.rivalries = tag.rivalries.filter((riv) => riv._id !== rivalry._id); // removes that rivalry from tag rivalries
+    //     tag.save();
+    //   }
+    // });
+    await rivalry.save();
 
     return res.send({ rivalry });
   } catch (err) {
@@ -230,6 +298,16 @@ router.put("/:rivalryId", authMiddleware, async (req, res) => {
 //delete
 router.delete("/:rivalryId", authMiddleware, async (req, res) => {
   try {
+    let rivalry = await Rivalry.findById(req.params.rivalryId);
+    console.log(req.userId);
+    console.log(rivalry.user);
+    if (req.userId.toString() !== rivalry.user.toString()) {
+      //user does not own the rivalry, so don't delete
+      return res
+        .status(400)
+        .send({ error: "Only the creator can delete the rivalry" });
+    }
+
     await Rivalry.findByIdAndRemove(req.params.rivalryId);
 
     return res.send({ msg: "Rivalry deleted" });
